Write shell history before entering config mode
On DANOS shell history entries are typically written to the history
file as soon as the corresponding command completes. This is done
by setting PROMPT_COMMAND to `history -a`, causing new history
entries to be written just before the prompt returns.

When entering configuration mode from operational mode via `configure`
a new shell is spawned, therefore the operational mode prompt does not
return until configuration mode is exited. This means that the history
entry for `configure` is written to the history file after any
commands entered in configuration mode. Although the timestamp of the
entry is correct this ordering can cause confusion.

To fix this add a call to `history -a` before entering configuration
mode. This ensures that `configure` is recorded in the history file
before any commands issued in configuration mode.

VRVDR-50705