Add support for decoding MgmtError.
Closes VRVDR-36175.

This commit provides support for decoding management errors in response to requests over the configd socket to perform actions such as validate or commit.  These are enhanced errors that contain fields mapping to the NETCONF (RFC 6241) <rpc-error> and which are represented inside configd by the (golang) MgmtError object.  These are encoded over the configd socket using JSON (though this will change to RFC 7951 in due course).  Multiple errors may be passed back from configd (eg for multiple validation errors).

While it's a large commit, it developed somewhat organically, and I'm not sure trying to split it into parts will add much.  Instead, here's a summary of the key changes:

The response struct may now contain a (decoded) mgmt_err_list struct, which is populated if the incoming JSON has a non-null 'mgmterror' field in it.  error_set_from_mgmt_error_list() populates a configd_error object from the incoming error.  The configd_error object is extended to include a list of management errors.  For backwards compatibility (and to avoid the need to change APIs), the existing 'text' field of the configd_error object is populated with either the message field (if a single management error) or a newline-separated string listing each message in a list of errors.

Further to this, commit and validate APIs now format the text message to mimic what was previously sent over the configd socket.  We no longer format at source because it is preferable to pass the error fields intact, and let the receiver format as desired.  The formatting should replicated as closely as possible the previously sent format.  Format calls are in transaction.c, with the code to do the formatting in error.c

Support for decoding management errors, and converting them to configd_error format, is added to error.c.

I've added cpputest-based unit tests, with the directory and makefile structure copied from another repo that also uses automake (so this is different to the setup in, for example, netconfd unit tests).  I used gcov to check that most of the new code was being tested, though didn't leave support in the makefile as it needed manual processing that didn't seem easy to integrate.

As well as the unit test, I have written acceptance tests that verify (for validate, edit-config, commit, get and get-config operations) that the expected <rpc-error> fields are passed from configd over the socket through the vyatta-cfg configd library and over NETCONF to the NETCONF client.

Additionally I have asked QA to run basic sanity on an ISO with all the changes here together with those in netconfd, configd, mgmterror and yang repos.  There was an initial crash due to some of my debug code assuming libjansson was slightly more robust than it seems to be (it has precious little checking for null pointers, and badly formatted JSON can send it into a permanent loop).  As this debug would have printed out log messages where previously we would not have had any logs I removed it. I'd thought it might be helpful to log errors when the caller didn't provide a configd_error object into which to dump the error, but figure that it's actually the caller saying 'dont' care' and thus we shouldn't log this.

Awaiting full regression results, but initail rerun with fixes looked promising ...