ifstats: Don't display unconfigured tunnel default
When the ip_gre module is loaded, two default tunnel interfaces are
created: gre0 and erspan0. gre0 is not a support interface and is
already ignored by "show interfaces.". However, erspan0 could be
a configured interface, so it is displayed, even if not configured.
This is possibly confusing to the end user.

If we find an erspan0 during the interface walk, check to see
if the interface is configured in the system, otherwise skip it.

Bugfix: VRVDR-43463