Move vyatta-interfaces-v1.yang to vyatta-interfaces
At one point vyatta-cfg was a single large package that installed all
the YANG files. The first part of the work to fix that split that
package into smaller ones. The second half of this work is moving
those small packages to locations that make more sense.

vyatta-interfaces-v1.yang (and associated helper utilities) is a better
(and more logical) fit in the vyatta-interfaces repository.

Bugfix: VRVDR-51933